Add Dandelion monitor and remove transactions from stempool
