+++ grin/src/server.rs
@@ -26,6 +26,7 @@ use adapters::*;
 use api;
 use chain;
 use core::{genesis, global};
+use dandelion_monitor;
 use miner;
 use p2p;
 use pool;
@@ -77,7 +78,7 @@ impl Server {
 			pool_adapter.clone(),
 			pool_net_adapter.clone(),
 		)));
-		tx_pool.monitor_transactions(tx_pool, stop.clone());
+		dandelion_monitor::monitor_transactions(config.pool_config.clone(), tx_pool.clone(), stop.clone());
 
 		let chain_adapter = Arc::new(ChainToPoolAndNetAdapter::new(tx_pool.clone()));
